+++ src/tests/blockFrequency.c
@@ -732,7 +732,7 @@ BlockFrequency_metric_print(struct state *state, long int sampleCount, long int 
 	 * Compute uniformity p-value
 	 */
 	chi2 = 0.0;
-	expCount = sampleCount / state->tp.uniformity_bins;
+	expCount = (double)sampleCount / state->tp.uniformity_bins;
 	if (expCount <= 0.0) {
 		// Not enough samples for uniformity check
 		uniformity = 0.0;
