+++ src/tests/randomExcursionsVariant.c
@@ -986,7 +986,7 @@ RandomExcursionsVariant_metric_print(struct state *state, long int sampleCount, 
 	 * Compute uniformity p-value
 	 */
 	chi2 = 0.0;
-	expCount = sampleCount / state->tp.uniformity_bins;
+	expCount = (double)sampleCount / state->tp.uniformity_bins;
 	if (expCount <= 0.0) {
 		// Not enough samples for uniformity check
 		uniformity = 0.0;
